Code is cleaner, simplified and unnessarry code removed
