+++ src/main/java/frc/robot/commands/AlignClimber.java
@@ -8,10 +8,13 @@ import frc.robot.subsystems.ColorSensor;
 import frc.robot.subsystems.DriveSubsystem;
 
 public class AlignClimber extends CommandBase {
+  
   ColorSensor colorSensor;
   DriveSubsystem driveSubsystem;
 
-  /** Creates a new AlignClimber. */
+  private Color tapeBlack = new Color(0.259,0.473,0.268);
+  private Color carpetGrey = new Color(0.263,0.478,0.259);
+
   public AlignClimber(ColorSensor colorSensor, DriveSubsystem driveSubsystem) {
     this.colorSensor = colorSensor;
     this.driveSubsystem = driveSubsystem;
@@ -19,9 +22,6 @@ public class AlignClimber extends CommandBase {
     addRequirements(colorSensor);
     addRequirements(driveSubsystem);
   }
-
-  private Color tapeBlack = new Color(0.259,0.473,0.268);
-  private Color carpetGrey = new Color(0.263,0.478,0.259);
 
   // Called when the command is initially scheduled.
   @Override
@@ -41,14 +41,4 @@ public class AlignClimber extends CommandBase {
     if (match.color == tapeBlack) this.cancel();
     else driveSubsystem.tankDrive(driveSpeed, -driveSpeed);
   } 
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {}
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return false;
-  }
 }
