--- src/main/java/frc/robot/commands/ClearDriveData.java
@@ -1,30 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-/*
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.DriveSubsystem;
-
-// NOTE:  Consider using this command inline, rather than writing a subclass.  For more
-// information, see:
-// https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class ClearDriveData extends CommandBase {
-  DriveSubsystem driveSubsystem;
-  public ClearDriveData(DriveSubsystem driveSubsystem) {
-    // Use addRequirements() here to declare subsystem dependencies.
-    this.driveSubsystem = driveSubsystem;
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {}
-
-  @Override
-  public void execute() {
-    driveSubsystem.zeroHeading();
-    driveSubsystem.resetDriveEncoders();
-  }
-}
-*/
