--- src/main/java/frc/robot/commands/CommandGroup.java
@@ -1,18 +1,0 @@
-/*package frc.robot.commands;
-import edu.wpi.first.wpilibj.XboxController;
-import edu.wpi.first.wpilibj.GenericHID.Hand;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.DriveSubsystem;
-import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-public class CommandGroup {
-    
-   
-    /** Autonomous Mode Delayed ******************************************************
-     * An autonomous routine for driving forwards off the initiation line and then shooting the three pre-loaded balls after waiting a certain number of seconds */
-   /* public class AutonomousMode_Delayed extends SequentialCommandGroup {
- 
-      
-    }
-}
-*/
