+++ src/main/java/frc/robot/commands/ToggleSlow.java
@@ -4,27 +4,32 @@
 
 package frc.robot.commands;
 
-import edu.wpi.first.wpilibj2.command.InstantCommand;
+import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.DriveSubsystem;
 
-// NOTE:  Consider using this command inline, rather than writing a subclass.  For more
-// information, see:
-// https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class SlowOn extends InstantCommand {
+public class ToggleSlow extends CommandBase {
+  DriveSubsystem driveSubsystem;
 
-  private DriveSubsystem driveSubsystem;
-
-  public SlowOn(DriveSubsystem driveSubsystem) {
+  public ToggleSlow(DriveSubsystem driveSubsystem) {
     this.driveSubsystem = driveSubsystem;
     addRequirements(driveSubsystem);
   }
 
-  // Called when the command is initially scheduled.
   @Override
   public void initialize() {}
 
   @Override
   public void execute() {
     driveSubsystem.slowOn();
+  }
+
+  @Override
+  public void end(boolean interrupted) {
+    driveSubsystem.slowOff();
+  }
+
+  @Override
+  public boolean isFinished() {
+    return false;
   }
 }
