+++ src/main/java/frc/robot/commands/ToggleTurbo.java
@@ -4,17 +4,14 @@
 
 package frc.robot.commands;
 
-import edu.wpi.first.wpilibj2.command.InstantCommand;
+import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.DriveSubsystem;
 
-// NOTE:  Consider using this command inline, rather than writing a subclass.  For more
-// information, see:
-// https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class TurboOn extends InstantCommand {
+public class ToggleTurbo extends CommandBase {
 
-  private DriveSubsystem driveSubsystem;
+  DriveSubsystem driveSubsystem;
 
-  public TurboOn(DriveSubsystem driveSubsystem) {
+  public ToggleTurbo(DriveSubsystem driveSubsystem) {
     this.driveSubsystem = driveSubsystem;
     addRequirements(driveSubsystem);
   }
@@ -22,5 +19,10 @@ public class TurboOn extends InstantCommand {
   @Override
   public void execute() {
     driveSubsystem.turboOn();
+  }
+
+  @Override
+  public void end(boolean interrupted) {
+    driveSubsystem.turboOff();
   }
 }
